Ship part shop is now fleet-compatible

+++ ChartedWaters/State_showskills.cpp
@@ -37,7 +37,7 @@ void State_ShowSkills::redraw()
             continue;
         std::string name = it->first.substr(0, 7);
         if (name.size() < 7)
-        for (int i = 0; i <= 7 - name.size(); i++)
+        for (size_t i = 0; i <= 7 - name.size(); i++)
             name += " ";
         int level = it->second.getLevel();
         int points = it->second.netPointsNeeded();
